+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/governance-core/auth/src/main/java/com.alibaba.cloud.governance/auth/rules/OrRule.java
@@ -1,4 +1,4 @@
-package com.alibaba.cloud.governance.common.rules;
+package com.alibaba.cloud.governance.auth.rules;
 
 import java.util.List;
