+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/governance-core/auth/src/main/java/com.alibaba.cloud.governance/auth/rules/manager/HeaderRuleManager.java
@@ -1,7 +1,7 @@
-package com.alibaba.cloud.governance.common.rules.manager;
+package com.alibaba.cloud.governance.auth.rules.manager;
 
-import com.alibaba.cloud.governance.common.rules.auth.HttpHeaderRule;
-import com.alibaba.cloud.governance.common.rules.util.HeaderMatchUtil;
+import com.alibaba.cloud.governance.auth.rules.util.HeaderMatchUtil;
+import com.alibaba.cloud.governance.auth.rules.auth.HttpHeaderRule;
 import org.springframework.http.HttpHeaders;
 
 import java.util.Map;
