+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/resource-transform/pom.xml
@@ -15,7 +15,6 @@
 
     <modules>
         <module>istio-resource-transform</module>
-        <module>resource-transform-common</module>
     </modules>
 
     <dependencies>
@@ -26,6 +25,12 @@
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.alibaba.cloud</groupId>
+            <artifactId>auth</artifactId>
+            <version>2.2.9-SNAPSHOT</version>
+            <scope>compile</scope>
         </dependency>
     </dependencies>
