--- spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-governance/resource-transform/resource-transform-common/pom.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>resource-transform</artifactId>
-        <groupId>com.alibaba.cloud</groupId>
-        <version>${revision}</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>resource-transform-common</artifactId>
-    <name>resource-transform-common</name>
-    <dependencies>
-        <dependency>
-            <groupId>io.envoyproxy.controlplane</groupId>
-            <artifactId>api</artifactId>
-            <version>0.1.32</version>
-            <scope>compile</scope>
-        </dependency>
-    </dependencies>
-
-    <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
-    </properties>
-
-</project>
